refactor: extract the LTR-559 drive policy out of poly_keymap.c
The auto-brightness + idle-inhibit POLICY (the USER_SYNC_SLAVE_DATA pull
channel, the lux -> contrast sqrt curve, and the proximity wake) moves
verbatim into ltr559_policy.c/.h — ~210 lines out of the 5,124-line
poly_keymap.c, the first of the gradual extractions deferred from #236.

The split handler stays private to the new file; poly_keymap.c registers
it through poly_ltr559_register_split_handler() alongside the other
transaction registrations, and housekeeping keeps calling
poly_ltr559_drive() under the same COMMUNITY_MODULE + DRIVE gates. The
sensor telemetry heartbeat stays in housekeeping (it is a housekeeping
log line, not policy).

One latent nesting straightened out: the POLY_DUMMY_TXN_TEST no-op
handler sat INSIDE the POLYKYBD_LTR559_DRIVE block while its
registration had its own independent #ifdef, so the dummy-test config
without the drive policy could never have linked. The handler now sits
under its own #ifdef only. No shipping config changes: both variants
define POLYKYBD_LTR559_DRIVE and neither defines POLY_DUMMY_TXN_TEST.
